Frontier: Mostly decouple game from UI
This allows the game to support multiple UIs, including a graphical UI,
without having to rewrite any of the game's code.

The decoupling is almost done. There is still work to be done like
bounds checking that should be handled by the Game class, but I wanted
to make a commit with a workable version finally.