Frontier: Add highscores and the option to play again
scanner.nextLine() has to be called after every usage of
scanner.nextInt() in order to remove the newline from the buffer and
allow main() to use scanner.nextLine(). The input system is not very
robust or well-designed, but fixing it is low priority since it is not
very interesting/challenging.